Eliminate nil map entries on Merkle trie cache (#1372)
The Merkle trie was using the pageToNIDsPtr map to associate in-memory cached pages of nodes with the actual nodes. During trie rearrangements, nodes are being deleted, which could lead to deletion of entries from this map. The map was correctly deleting the internal list object, but left it as "nil" instead of completely eliminating it from the map. Over time, this could result is a memory leak.

The solution is quite straight-forward. Instead of just placing nil value on the map when the associated page gets empty, delete the page completely.